vertex-ar/database.py
import os
from datetime import datetime

from dotenv import load_dotenv
from sqlalchemy import Column, DateTime, Integer, String, create_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")
engine = create_engine(DATABASE_URL)
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

Base = declarative_base()


class AREntry(Base):
    __tablename__ = "ar_entries"

    id = Column(Integer, primary_key=True, index=True)
    uuid = Column(String, unique=True, index=True)
    created_at = Column(DateTime, default=datetime.utcnow)
    image_key = Column(String, nullable=False)
    video_key = Column(String, nullable=False)
    nft_prefix = Column(String, nullable=False)
    status = Column(String, default="active")


# Создание таблиц
Base.metadata.create_all(bind=engine)


def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()
