vertex-ar/requirements-simple.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
sqlite3
